# api/openapi.yaml
openapi: "3.0.3"
info:
  title: Potoo API
  version: "0.1.0"
  description: |
    Developer-first notification orchestration platform.

    Authenticate all requests with an API key in the `Authorization` header:

    ```
    Authorization: Bearer <api_key>
    ```
servers:
  - url: http://localhost:8080
    description: Local development
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: API key issued from the Potoo dashboard.
  schemas:
    AuthUser:
      type: object
      required: [id, email, created_at]
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        created_at:
          type: string
          format: date-time
...
    ApiKey:
      type: object
      required: [id, name, key_prefix, scopes, created_at]
      properties:
        id:
          type: string
        name:
          type: string
        key_prefix:
          type: string
          description: First characters of the key — safe to display, not secret
        scopes:
          type: array
          items:
            type: string
        last_used_at:
          type: string
          format: date-time
        revoked_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    CreateAPIKeyRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: Production key
        scopes:
          type: array
          items:
            type: string
          default: []
    CreateAPIKeyResponse:
      allOf:
        - $ref: "#/components/schemas/ApiKey"
        - type: object
          required: [key]
          properties:
            key:
              type: string
              description: Full API key — shown once, never retrievable again
...
    InviteAcceptRequest:
      type: object
      required: [token, first_name, last_name, email, password]
      properties:
        token:
          type: string
        first_name:
          type: string
          example: James
        last_name:
          type: string
          example: Zaman
        phone:
          type: string
          example: "+447911123456"
        email:
          type: string
          format: email
          example: james@example.com
        password:
          type: string
          minLength: 8
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
              example: validation_failed
            message:
              type: string
              example: "Template is missing required variable: first_name"
            request_id:
              type: string
              example: req_01HX1234
            docs_url:
              type: string
              example: https://docs.potoo.com/errors/validation_failed
...
security:
  - ApiKeyAuth: []
paths:
...
  /v1/auth/accept-invite:
    post:
      operationId: acceptInvite
      summary: Accept an invite and create a user account
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteAcceptRequest"
      responses:
        "201":
          description: Account created and joined organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "410":
          description: Invite already used or expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /v1/auth/refresh:
    post:
      operationId: refreshSession
      summary: Exchange a refresh token for a new access + refresh token pair
      tags: [auth]
      security: []
      responses:
        "200":
          description: New tokens issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /v1/api-keys:
    get:
      operationId: listAPIKeys
      summary: List API keys for the session org
      tags: [api-keys]
      security: []
      responses:
        "200":
          description: Key list
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ApiKey"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: createAPIKey
      summary: Create an API key
      tags: [api-keys]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAPIKeyRequest"
      responses:
        "201":
          description: Key created — full key shown once
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateAPIKeyResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not authorised to create a key for that org
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/api-keys/{keyId}:
    delete:
      operationId: revokeAPIKey
      summary: Revoke an API key
      tags: [api-keys]
      security: []
      parameters:
        - name: keyId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Key revoked
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Key not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...